Move utils to ArrayHelper.cs
