+++ source/Cosmos.System2/Network/IPv4/TCP/Tcp.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using Cosmos.HAL;
+using Cosmos.System.Helpers;
 using Cosmos.System.Network.Config;
 
 namespace Cosmos.System.Network.IPv4.TCP
@@ -282,7 +283,7 @@ namespace Cosmos.System.Network.IPv4.TCP
 
                         LastSequenceNumber = packet.SequenceNumber;
 
-                        data = Concat(data, packet.TCP_Data);
+                        data = ArrayHelper.Concat(data, packet.TCP_Data);
 
                         rxBuffer.Enqueue(packet);
 
@@ -303,7 +304,7 @@ namespace Cosmos.System.Network.IPv4.TCP
                     {
                         AckNumber += packet.TCP_DataLength;
 
-                        data = Concat(data, packet.TCP_Data);
+                        data = ArrayHelper.Concat(data, packet.TCP_Data);
                     }
                 }
                 return;
@@ -464,63 +465,6 @@ namespace Cosmos.System.Network.IPv4.TCP
 
             OutgoingBuffer.AddPacket(packet);
             NetworkStack.Update();
-        }
-
-        /// <summary>
-        /// Contatenate two byte arrays. https://stackoverflow.com/a/45531730
-        /// </summary>
-        /// <param name="first">First byte array.</param>
-        /// <param name="second">Byte array to concatenate.</param>
-        private byte[] Concat(byte[] first, byte[] second)
-        {
-            byte[] output;
-            int alen = 0;
-
-            if (first == null)
-            {
-                output = new byte[second.Length];
-            }
-            else
-            {
-                output = new byte[first.Length + second.Length];
-                alen = first.Length;
-                for (int i = 0; i < first.Length; i++)
-                {
-                    output[i] = first[i];
-                }
-            }
-            for (int j = 0; j < second.Length; j++)
-            {
-                output[alen + j] = second[j];
-            }
-            return output;
-        }
-
-        /// <summary>
-        /// Split byte array into chunks of a specified size.
-        /// </summary>
-        /// <param name="buffer">Byte array to split.</param>
-        /// <param name="chunksize">Chunk size.</param>
-        internal byte[][] ArraySplit(byte[] buffer, int chunksize = 1000)
-        {
-            var size = buffer.Length;
-            var chunkCount = (buffer.Length + chunksize - 1) / chunksize;
-            var bufferArray = new byte[chunkCount][];
-            int index = 0;
-
-            for (var i = 0; i < chunkCount; i++)
-            {
-                bufferArray[i] = new byte[Math.Min(chunksize, buffer.Length - i * chunksize)];
-            }
-            for (var i = 0; i < chunkCount; i++)
-            {
-                for (var j = 0; j < bufferArray[i].Length; j++)
-                {
-                    bufferArray[i][j] = buffer[index];
-                    index++;
-                }
-            }
-            return bufferArray;
         }
 
         #endregion
